src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/content-pseudo-element-overridden-2-expected.txt

CONSOLE ERROR: line 87: Uncaught TypeError: element.createShadowRoot is not a function
Test for crbug.com/274059. Compare rules from a style in a shadow tree with ::content in a different shadow tree.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
